SonarCube 対応
class を別ファイルに分離
